requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.4.0
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.11.0
APScheduler==3.10.4
attrs==25.4.0
bcrypt==5.0.0
black==24.10.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.0
frozenlist==1.8.0
google-ads==26.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.1.3
oauthlib==3.3.1
openai==1.54.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.5
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
